working on main URL redirect functionality

+++ URL-Shortener/API/routes/user.py
@@ -2,7 +2,6 @@ from API.db import User, get_async_session
 from API.exceptions import (
     IntegrityError,
     UserAlreadyExists,
-    UserNotFound,
     WrongPassword,
 )
 from API.schemas import (
@@ -13,7 +12,7 @@ from API.schemas import (
     UserUpdateSchema,
 )
 from API.services import UserService, validate_token
-from fastapi import APIRouter, Depends, status
+from fastapi import APIRouter, Depends, status, HTTPException
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 user_router = APIRouter()
@@ -40,9 +39,9 @@ async def create_user(
             email=user.email,
         )
     except UserAlreadyExists as exp:
-        raise exp.to_http()
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=f"{exp.email} {exp.message}")
     except IntegrityError as exp:
-        raise exp.to_http()
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="Intern db error")
 
 
 @user_router.get(
@@ -56,9 +55,11 @@ async def get_current_user(
     token: TokenPayload = Depends(validate_token),
     session: AsyncSession = Depends(get_async_session),
 ) -> UserResponseSchema:
+    if token is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authenticated")
     user: User | None = await user_service.get_user_by_id(token.user_id, session)
     if not user:
-        raise UserNotFound().to_http()
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
     return UserResponseSchema(
         id=user.id,
         created_at=user.created_at,
@@ -80,9 +81,11 @@ async def update_current_user(
     token: TokenPayload = Depends(validate_token),
     session: AsyncSession = Depends(get_async_session),
 ) -> UserResponseSchema:
+    if token is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authenticated")
     user: User | None = await user_service.get_user_by_id(token.user_id, session)
     if not user:
-        raise UserNotFound().to_http()
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
     try:
         updated_user = await user_service.update_user(user, user_data, session)
         return UserResponseSchema(
@@ -93,9 +96,9 @@ async def update_current_user(
             email=updated_user.email,
         )
     except WrongPassword as exp:
-        raise exp.to_http()
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=exp.message)
     except IntegrityError as exp:
-        raise exp.to_http()
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="Intern db error")
 
 
 @user_router.delete(
@@ -109,10 +112,14 @@ async def delete_current_user(
     token: TokenPayload = Depends(validate_token),
     session: AsyncSession = Depends(get_async_session),
 ) -> None:
+    if token is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authenticated")
     user: User | None = await user_service.get_user_by_id(token.user_id, session)
     if not user:
-        raise UserNotFound().to_http()
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
     try:
         await user_service.delete_user(user_data, user, session)
     except WrongPassword as exp:
-        raise exp.to_http()
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=exp.message)
+    except IntegrityError as exp:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="Intern db error")
